Rename: store.js 이름 변경 - store는 1개만 있는 경우가 일반적이므로 통합적인 내용음 담고자 store.js   로 변경
