+++ src/main.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.jsx';
 import { Provider } from 'react-redux';
-import store from './redux/store/postStore';
+import App from './App.jsx';
+import store from './redux/store/store';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
